requirements.txt
docker==7.0.0
packaging==23.2
python-dotenv==1.0.1
slack_sdk==3.27.1